A bit more cleanup
